@@ -4,15 +4,14 @@ import io.getunleash.Unleash;
 import io.getunleash.util.UnleashConfig;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
-import javax.annotation.PreDestroy;
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import jakarta.annotation.PreDestroy;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import java.net.URISyntaxException;
-import java.util.Random;
 
 @Path("/")
 public class DemoApp {
